+++ Kernel/Output/HTML/Dashboard/TicketGeneric.pm
@@ -1750,7 +1750,7 @@ sub Run {
                         Age                => $EscalationData{EscalationTime},
                         TimeShowAlwaysLong => 1,
                         Space              => ' ',
-                    );
+                    ) || '-';
                     $EscalationData{EscalationTimeWorkingTime} = $LayoutObject->CustomerAge(
                         Age                => $EscalationData{EscalationTimeWorkingTime},
                         TimeShowAlwaysLong => 1,
@@ -1886,7 +1886,7 @@ sub Run {
                     $LayoutObject->Block(
                         Name => "ContentLargeTicketGenericColumn$BlockType",
                         Data => {
-                            GenericValue => $DataValue || '',
+                            GenericValue => $DataValue || '-',
                             Class        => $CSSClass  || '',
                         },
                     );