+++ Kernel/Output/HTML/TicketOverview/Small.pm
@@ -1515,7 +1515,7 @@ sub Run {
                         Age                => $EscalationData{EscalationTime},
                         TimeShowAlwaysLong => 1,
                         Space              => ' ',
-                    );
+                    ) || '-';
                     $EscalationData{EscalationTimeWorkingTime} = $LayoutObject->CustomerAge(
                         Age                => $EscalationData{EscalationTimeWorkingTime},
                         TimeShowAlwaysLong => 1,
@@ -1614,7 +1614,7 @@ sub Run {
                 $LayoutObject->Block(
                     Name => "RecordTicketColumn$BlockType",
                     Data => {
-                        GenericValue => $DataValue || '',
+                        GenericValue => $DataValue || '-',
                         Class        => $CSSClass  || '',
                     },
                 );